requirements.txt
Flask
opencv-python
opencv-contrib-python  
numpy
requests
Pillow
tk